# Make 64-bit printing on 32-bit platforms 9% to 12% faster.