Printing 64-bit doubles on 32-bit platforms is the slowest scenario, but there's a simple way to improve this. Because the core Ryu algorithm is so fast, a relatively large amount of time is spent on the mundane task of decomposing an integer into decimal digits. This involves expensive 64-bit divisions, partially mitigated by the digit table. It's possible to do most of the work with cheap 32-bit divisions. The strategy is:

* When we detect a 32-bit platform, activate the new code. (I'm currently checking MSVC's platform macros, which I'm familiar with. Other platform macros can be added in the future. Note that I've phrased this preprocessor check so that unknown platforms use the original code. Alternatively, the preprocessor could be avoided by testing `sizeof(void *)` and using `size_t` for the type of `output2`, although a "conditional expression is constant" warning would need to be silenced.)

* Loop while `output` doesn't fit within `uint32_t`. (Tested by examining the high 32 bits, in an attempt to be compiler-friendly.)

* Perform an expensive 64-bit division by 10^9. (This loop will run at most twice, for 19-digit and 20-digit numbers.)

* Use 32-bit divisions to print 9 digits at a time. Note that this prints leading zeroes because we know we have more work to do. (This loop runs when `output` is at least 2^32, so `output` has at least 10 digits.)

* When `output` fits within `uint32_t`, we exit the loop, assign it to `output2`, and perform the rest of the work with that 32-bit value.

* For other platforms, we define `output2` to be `uint64_t`, allowing the following code to be shared.

I observe that this improves performance for both MSVC and Clang 6.0.0. Tested with /MT /O2 but without link-time codegen:

MSVC before: 202.369 ns
MSVC after: 185.923 ns
Clang before: 149.431 ns
Clang after: 133.485 ns

This also contains a minor improvement to remove unnecessary loops. After `while (output >= 10000)` terminates, output is at most 9999. The `(output >= 100)` loop contains `output /= 100;`, so after one iteration, output is at most 99, so there can't be a second iteration. Therefore, this can just be an if-statement, slightly improving codegen.